+++ packages/sample/src/application.ts
@@ -5,6 +5,7 @@ import {
   pipeline,
   projectionViewMatrix,
   sprite,
+  tilingSprite,
 } from '@story75/pixel-forge';
 import Stats from 'stats.js';
 
@@ -12,74 +13,74 @@ export async function application(canvas: HTMLCanvasElement): Promise<void> {
   const stats = new Stats();
   document.body.appendChild(stats.dom);
 
-  const { device, context } = await createContext(canvas);
+  const context = await createContext(canvas);
 
   const projectionViewMatrixUniformBuffer = projectionViewMatrix(
-    device,
+    context.device,
     canvas.width,
     canvas.height,
   );
 
-  const textureLoader = createTextureLoader(device);
+  const textureLoader = createTextureLoader(context.device);
   const tex = await textureLoader('assets/pixel-prowlers.png');
 
-  const renderPass = pipeline(
-    device,
-    context,
-    projectionViewMatrixUniformBuffer,
-  );
+  const renderPass = pipeline(context, projectionViewMatrixUniformBuffer);
 
-  const sprites: Sprite[] = [];
+  const [x, y] = [
+    canvas.width / 2 - tex.width / 2,
+    canvas.height / 2 - tex.height / 2,
+  ];
 
-  sprites.push(
+  const rotatingSprites: Sprite[] = [
     sprite({
       texture: tex,
-      x: 300,
-      y: 300,
+      x,
+      y,
       origin: [0, 0],
     }),
-  );
-  sprites.push(
     sprite({
       texture: tex,
-      x: 300,
-      y: 300,
+      x,
+      y,
       origin: [0, 1],
     }),
-  );
-  sprites.push(
     sprite({
       texture: tex,
-      x: 300,
-      y: 300,
+      x,
+      y,
       origin: [1, 0],
     }),
-  );
-  sprites.push(
     sprite({
       texture: tex,
-      x: 300,
-      y: 300,
+      x,
+      y,
       origin: [1, 1],
     }),
-  );
-  sprites.push(
     sprite({
       texture: tex,
-      x: 300,
-      y: 300,
+      x,
+      y,
       origin: [0.5, 0.5],
     }),
-  );
+  ];
+
+  const parallax = tilingSprite({
+    texture: tex,
+    width: canvas.width,
+    height: canvas.height,
+  });
+
+  parallax.offset[0] = 1.5;
+  parallax.offset[1] = 1.5;
 
   const draw = function () {
     stats.begin();
 
-    for (const sprite of sprites) {
+    for (const sprite of rotatingSprites) {
       sprite.rotation += 0.01;
     }
 
-    renderPass(sprites);
+    renderPass([parallax, ...rotatingSprites]);
 
     stats.end();
     requestAnimationFrame(draw);